+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Clear natural healing status effect when BODY returns to max. [#686](https://github.com/dmdorman/hero6e-foundryvtt/issues/686)
 - Improve AOE modifier advantage calculation and descriptions including 5e size calculations.
 - Movement powers now account for Reduced/Increased END modifiers. Still issues with movements with multiple movement types. [#599](https://github.com/dmdorman/hero6e-foundryvtt/issues/599)
+- Token selected movement is now updated for GM when players make a change via movement radio controls. [#320](https://github.com/dmdorman/hero6e-foundryvtt/issues/320)
 
 ## Version 3.0.57